requirements.txt
streamlit
opencv-contrib-python
pandas
numpy
Pillow
streamlit-option-menu
streamlit-webrtc
